set 1 challenge 6: Take first 4 blocks, not first 2
As step 4 suggests, it would be wise to take the average (actually, sum
is sufficient) of the first four blocks (I do blocks 0+1, blocks 1+2,
blocks 2+3).

This makes the correct keysize appear much earlier in the list.

Failure to do this doesn't impede correctness (using 2 blocks, the
original keysize is the first keysize that produces a readable answer,
so it eventually gets found), but causes more work to be done.